Write a script that runs a C file through the preprocessor and save the result into another file.
The C file name will be saved in the variable main.c
The output should be saved in the file c
Write a script that compiles a C file but does not link.

The C file name will be saved in the variable main.c
The output file should be named the same as the C file, but with the extension .o instead of .c.
Example: if the C file is main.c, the output file should be main.o